+++ .gitmodules
@@ -1,0 +1,3 @@
+[submodule "genepy/epigenetics/rose"]
+	path = genepy/epigenetics/rose
+	url = https://github.com/jkobject/rose.git
