+++ genepy/epigenetics/chipseq.py
@@ -33,6 +33,18 @@ chroms = {'chr1', 'chr10', 'chr11', 'chr12', 'chr13', 'chr14', 'chr15', 'chr16',
 def bigWigFrom(bams, folder="", numthreads=8, genome='GRCh37', scaling=None, verbose=1):
 	"""
 	run the bigwig command line for a set of bam files in a folder
+
+	Can apply some scaling and process files in parallel
+
+	Args:
+	----
+		bams: list[str] of filepath to bam files
+		folder: str folder where the bam files would be stored (will save the bigwigs there as well, else to current path ./)
+		numthreadds: int number of threads to process in parallel
+		genome: str genome version to use (for genome size) only GRCh37 GRCh38 (see size global variable in this file)
+		scaling: list[float] a list of scaling values to apply to the bigwigs. if provided, won't scale to read counts.
+		verbose: int verbose level
+
 	"""
 	if "bigwig" not in os.listdir(folder if folder else '.'):
 		os.mkdir(folder + "bigwig")
@@ -51,6 +63,21 @@ def bigWigFrom(bams, folder="", numthreads=8, genome='GRCh37', scaling=None, ver
 
 
 def ReadRoseSuperEnhancers(roseFolder, containsINPUT=True, name="MV411"):
+	"""
+	reads ROSE2's output and returns its superenhancer bedfile as a pd dataframe. 
+
+	Can be multiple superenhancers from a set of HRK27ac bam files
+
+	Args:
+	-----
+		roseFolder: str folderpath to ROSE's output
+		containsINPUT: bool whether the bedfile contains INPUT signal as well
+		name: str sample name from which we computed superenhancers
+
+	Returns:
+	--------
+		a dataframe of the bed representation of the superenhancers
+	"""
 	beds = os.listdir(roseFolder)
 	superenhan = pd.DataFrame()
 	for i in beds:
@@ -66,13 +93,25 @@ def ReadRoseSuperEnhancers(roseFolder, containsINPUT=True, name="MV411"):
 	return superenhan.sort_values(by=['chrom','start','end'])
 
 
-def loadPeaks(peakFile=None,peakfolder=None, isMacs=True, CTFlist=[], skiprows=0):
-	# for each data peak type file listed in a MACS2 way in a given MACS2 output folder, will merge them
-	# all into one dataframe and output the dataframe
+def loadPeaks(peakFile=None, peakfolder=None, isNF=True, CTFlist=[], skiprows=0):
+ 	"""
+	loads 1 to many peak bedfile into one pandas dataframe.
+
+	all og the peaks will be concatenated into one dataframe. this function can 
+	work with jkobject|nfcore/chipseq nextflow pipelines output (isNF)
+
+	Args:
+	-----
+		peakFile: str filepath to a peak bedfile
+		peakfolder: str folderpath to a folder congaining beddfile
+		isNF
+		CTFlist
+		skiprows
+	"""
 	if peakfolder:
 		bindings = pd.DataFrame()
 		for folder in os.listdir(peakfolder):
-			if isMacs:
+			if isNF:
 				if any(tf in folder for tf in CTFlist) or not CTFlist:
 					binding = pd.read_csv(peakfolder + folder + "/NA_peaks.narrowPeak", sep='\t', header=None) if\
 					os.exists(peakfolder + folder + "/NA_peaks.narrowPeak") else peakfolder + folder + "/NA_peaks.broadPeak"
@@ -115,260 +154,6 @@ def loadPeaks(peakFile=None,peakfolder=None, isMacs=True, CTFlist=[], skiprows=0
 	bindings.foldchange = bindings.foldchange.astype('float')
 	bindings = bindings.reset_index(drop=True)
 	return bindings
-
-
-def pysam_getPeaksAt(peaks, bams, folder='data/seqs/', window=1000, numpeaks=1000, numthreads=8):
-
-	# get pysam data
-	# ask for counts only at specific locus based on windows from center+-size from sorted MYC peaks
-	# for each counts, do a rolling average (or a convolving of the data) with numpy
-	# append to an array
-	# return array, normalized
-	loaded = {}
-	res = {i: np.zeros((len(peaks), window * 2)) for i in bams}
-	peaks = peaks.sort_values(by="foldchange", ascending=False).iloc[:numpeaks]
-	peaks.chrom = peaks.chrom.astype(str)
-	for val in bams:
-		loaded.update({val: pysam.AlignmentFile(folder + val, 'rb', threads=numthreads)})
-	for k, bam in loaded.items():
-		for num, (i, val) in enumerate(peaks.iterrows()):
-			print(int(num / len(peaks)), end='\r')
-			center = int((val['start'] + val['end']) / 2)
-			for pileupcolumn in bam.pileup(val['chrom'], start=center - window,
-										   stop=center + window, truncate=True):
-				res[k][num][pileupcolumn.pos - (center - window)] = pileupcolumn.n
-	fig, ax = plt.subplots(1, len(res))
-	for i, (k, val) in enumerate(res.items()):
-		sns.heatmap(val, ax=ax[i])
-		ax[i].set_title(k.split('.')[0])
-	fig.show()
-	return res, fig
-
-
-def bedtools_getPeaksAt(peaks, bams, folder='data/seqs/', window=1000, numpeaks=1000, numthreads=8):
-	"""
-	get pysam data
-	ask for counts only at specific locus based on windows from center+-size from sorted MYC peaks
-	for each counts, do a rolling average (or a convolving of the data) with numpy
-	append to an array
-	return array, normalized
-	"""
-	loaded = {}
-	center = [int((val['start'] + val['end']) / 2) for k, val in peaks.iterrows()]
-	peaks['start'] = [c - window for c in center]
-	peaks['end'] = [c + window - 1 for c in center]
-	peaks[peaks.columns[:3]].sort_values(by=['chrom', 'start']).to_csv('temp/peaks.bed', sep='\t', index=False, header=False)
-	bedpeaks = BedTool('temp/peaks.bed')
-
-	fig, ax = plt.subplots(1, len(bams))
-	peakset = peaks["foldchange"].values.argsort()[::-1][:numpeaks]
-	for i, val in enumerate(bams):
-		coverage = BedTool(folder + val).intersect(bedpeaks).genome_coverage(bga=True, split=True)\
-			.intersect(bedpeaks).to_dataframe(names=['chrom', 'start', 'end', 'coverage'])
-		cov = np.zeros((len(peaks), window * 2), dtype=int)
-		j = 0
-		pdb.set_trace()
-		for i, (k, val) in enumerate(peaks.iterrows()):
-			print(i / len(peaks), end='\r')
-			while coverage.iloc[j].start > val.start:
-				j -= 1
-			while coverage.iloc[j].start < val.end:
-				cov[i][coverage.iloc[j].start - val.start:coverage.iloc[j].end - val.start] =\
-					coverage.iloc[j].coverage
-				j += 1
-		sns.heatmap(coverage, ax=ax[i])
-		ax[i].set_title(val.split('.')[0])
-	fig.show()
-	return None, fig
-
-
-def makeProfiles(matx=[], folder='', matnames=[], title='',
-				   name='temp/peaksat.pdf', refpoint="TSS", scale=None,
-				   sort=False, withDeeptools=True, cluster=1, vmax=None, vmin=None, overlap=False,
-				   legendLoc=None):
-	if withDeeptools:
-		if not (len(matnames) == 2 and len(matx) == 2):
-			raise ValueError('you need two mat.gz files and two names')
-		h.createFoldersFor(name)
-		cmd = 'computeMatrixOperations relabel -m '
-		cmd += matx[0] + ' -o '+matx[0]+' --groupLabels '+matnames[0]
-		cmd += ' && computeMatrixOperations relabel -m '
-		cmd += matx[1] + ' -o '+matx[1]+' --groupLabels '+matnames[1]
-		cmd += ' && computeMatrixOperations rbind -m '
-		cmd += matx[0] + ' ' + matx[1] + " -o " + '.'.join(name.split('.')[:-1]) + ".gz"
-		cmd += ' && plotProfile'
-		cmd += " --matrixFile " + '.'.join(name.split('.')[:-1]) + ".gz"
-		cmd += " --outFileName " + name
-		cmd += " --refPointLabel " + refpoint
-		if vmax is not None:
-			cmd += " -max "+str(vmax)
-		if vmin is not None:
-			cmd += " -min "+str(vmin)
-		if cluster >1:
-			cmd += " --perGroup --kmeans "+str(cluster)
-		if legendLoc:
-			cmd += " --legendLocation "+legendLoc
-		if title:
-			cmd += " --plotTitle " + title
-		data = subprocess.run(cmd, shell=True, capture_output=True)
-		print(data)
-
-def getPeaksAt(peaks, bigwigs, folder='', bigwignames=[], peaknames=[], window=1000, title='', numpeaks=4000, numthreads=8,
-				   width=5, length=10,torecompute=False, name='temp/peaksat.pdf', refpoint="TSS", scale=None,
-				   sort=False, withDeeptools=True, onlyProfile=False, cluster=1, vmax=None, vmin=None, overlap=False,
-				   legendLoc=None):
-	"""
-	get pysam data
-	ask for counts only at specific locus based on windows from center+-size from sorted MYC peaks
-	for each counts, do a rolling average (or a convolving of the data) with numpy
-	append to an array
-	return array, normalized
-	"""
-	if withDeeptools:
-		if isinstance(peaks, pd.DataFrame):
-			peaks = 'peaks.bed '
-			peaks.to_csv('peaks.bed', sep='\t', index=False, header=False)
-		elif type(peaks) == list:
-			pe = ''
-			i=0
-			for n, p in enumerate(peaks):
-				if 20 < int(os.popen('wc -l ' + p).read().split(' ')[0]):
-					pe += p + ' '
-				elif len(peaknames) > 0:
-					peaknames.pop(n-i)
-					i+=1
-			peaks = pe
-		elif type(peaks) == str:
-			peaks += ' '
-		else:
-			raise ValueError(' we dont know this filetype')
-		if type(bigwigs) is list:
-			pe = ''
-			for val in bigwigs:
-				pe += folder + val + ' '
-			bigwigs = pe
-		else:
-			bigwigs = folder + bigwigs + ' '
-		h.createFoldersFor(name)
-		cmd= ''
-		if not os.path.exists('.'.join(name.split('.')[:-1]) + ".gz") or torecompute:
-			cmd += "computeMatrix reference-point -S "
-			cmd += bigwigs
-			cmd += " --referencePoint "+refpoint
-			cmd += " --regionsFileName " + peaks
-			cmd += " --missingDataAsZero"
-			cmd += " --outFileName " + '.'.join(name.split('.')[:-1]) + ".gz"
-			cmd += " --upstream " + str(window) + " --downstream " + str(window)
-			cmd += " --numberOfProcessors " + str(numthreads) + ' && '
-		cmd += "plotHeatmap" if not onlyProfile else 'plotProfile'
-		if type(name) is list:
-			if not onlyProfile:
-				raise ValueError('needs to be set to True, can\'t average heatmaps')
-			cmd+= " --matrixFile " + '.gz '.join(name) + ".gz"
-			if average:
-				cmd+= "--averageType mean"
-		else:
-			cmd += " --matrixFile " + '.'.join(name.split('.')[:-1]) + ".gz"
-		cmd += " --outFileName " + name
-		cmd += " --refPointLabel "+ refpoint
-		if vmax is not None:
-			cmd += " -max "+str(vmax)
-		if vmin is not None:
-			cmd += " -min "+str(vmin)
-		if cluster>1:
-			cmd += " --perGroup --kmeans "+str(cluster)
-		if overlap:
-			if onlyProfile:
-				cmd += " --plotType overlapped_lines"
-			else:
-				raise ValueError("overlap only works when onlyProfile is set")
-		if legendLoc:
-			cmd+=" --legendLocation "+legendLoc
-
-		if len(peaknames) > 0:
-			pe = ''
-			for i in peaknames:
-				pe += ' ' + i
-			cmd += " --regionsLabel" + pe
-		if type(bigwigs) is list:
-			if len(bigwignames) > 0:
-				pe = ''
-				for i in bigwignames:
-					pe += ' "' + i +'"'
-				cmd += " --samplesLabel" + pe
-		if title:
-			cmd += " --plotTitle '"+title+"'"
-		data = subprocess.run(cmd, shell=True, capture_output=True)
-		print(data)
-	else:
-		if 'relative_summit_pos' in peaks.columns:
-			center = [int((val['start'] + val['relative_summit_pos'])) for k, val in peaks.iterrows()]
-		else:
-			center = [int((val['start'] + val['end']) / 2) for k, val in peaks.iterrows()]
-		pd.set_option('mode.chained_assignment', None)
-		peaks['start'] = [c - window for c in center]
-		peaks['end'] = [c + window for c in center]
-		fig, ax = plt.subplots(1, len(bigwigs), figsize=[width, length], title=title if title else 'Chip Heatmap')
-		if sort:
-			peaks = peaks.sort_values(by=["foldchange"], ascending=False)
-		if numpeaks > len(peaks):
-			numpeaks = len(peaks) - 1
-		cov = {}
-		maxs = []
-		for num, bigwig in enumerate(bigwigs):
-			bw = pyBigWig.open(folder + bigwig)
-			co = np.zeros((numpeaks, window * 2), dtype=int)
-			scale = scale[bigwig] if scale is dict else 1
-			for i, (k, val) in enumerate(peaks.iloc[:numpeaks].iterrows()):
-				try:
-					co[i] = np.nan_to_num(bw.values(str(val.chrom), val.start, val.end), 0)
-				except RuntimeError as e:
-					print(str(val.chrom), val.start, val.end)
-					pass
-			cov[bigwig] = co
-			maxs.append(co.max())
-		for num, bigwig in enumerate(bigwigs):
-			sns.heatmap(cov[bigwig] * scale, ax=ax[num], vmax=max(maxs), yticklabels=[], cmap=cmaps[num],
-						cbar=True)
-			ax[num].set_title(bigwig.split('.')[0])
-		fig.subplots_adjust(wspace=0.1)
-		fig.show()
-		fig.savefig(name)
-		return cov, fig
-
-
-"""
-def computeMeanCov(bigwigFolder, meanOnly=True, ref="GRCh38", averageFragSize=150, outputfolder='/data/coverage/'):
-	meancov = {}
-	for val in os.listdir(bigwigFolder):
-		bw = pyBigWig.open(folder + bigwig)
-		if bw:
-			meancov[val.split('.')[0]] = bw.
-	return meancov
-
-
-def substractPeaks(peaks1, to):
-	peaks1 = peaks1.sort_values(by=['chrom', 'start'])
-	peaks2 = to.sort_values(by=['chrom', 'start'])
-	j = 0
-	i = 0
-	newpeaks = pd.DataFrame(columns=peaks2.columns)
-	while i < len(peaks2):
-		if peaks1.iloc[j].chrom == peaks2.iloc[j].chrom:
-			else
-		intersection = h.intersection([peaks2.iloc[i]['start'], peaks2.iloc[i]['end']],
-										   [peaks1.iloc[j]['start'], peaks1.iloc[j]['start']])
-		if intersection:
-
-			j += 1
-		while peaks2.iloc[i]['end'] > peaks1.iloc[j]['start']:
-			if peaks2.iloc[i]['end'] > peaks1.iloc[j]['end']:
-				newpeaks.append(peaks2.iloc[i], ignore_index=True)
-			else:
-				newpeaks.append({'start': peaks2.iloc[i]['start'],
-								 'end': peaks1.iloc[j]['start'],
-								 'chromomsome': peaks1.iloc[j]['chrom']}, ignore_index=True)
-"""
 
 
 def simpleMergePeaks(peaks, window=0, totpeaknumber=0, maxp=True, mergedFold="mean"):
