+++ genepy/utils/plot.py
@@ -20,7 +20,6 @@ import bokeh
 import colorcet as cc
 from PIL import Image, ImageDraw, ImageFont
 import seaborn as sns
-from genepy.epigenetics import chipseq as chip 
 
 import pandas as pd
 from math import pi
@@ -613,40 +612,3 @@ def SOMPlot(net, size, colnames, minweight=0.1, distq1=0.535, distq2=0.055, dist
         somnodes.loc[i, 'features'] = tot
     #interactive SOM with features with highest importance to the nodes, displayed when hovering
     bigScatter(somnodes, precomputed=True, features=True, binsize=1, title='Cobinding SOM cluster of '+str(size), folder=folder)
-
-
-def andrew(groups, merged, annot, enr=None, pvals=None, cols=8, precise=True, title = "sorted clustermap of cobindings clustered", folder="", rangeval=4, okpval=10**-3, size=(20,15),vmax=3, vmin=0):
-    if enr is None or pvals is None:
-        enr, pvals = chip.enrichment(merged, groups=groups)
-    rand = np.random.choice(merged.index,5000)
-    subgroups = groups[rand]
-    sorting = np.argsort(subgroups)
-    redblue = cm.get_cmap('RdBu_r',256)
-    subenr = enr.iloc[annot-cols:]
-    subenr[subenr>rangeval]=rangeval
-    subenr[subenr<-rangeval]=-rangeval
-    subenr = subenr/rangeval
-    data = []
-    #colors = []
-    impv = pvals.values
-    for i in subgroups[sorting]:
-        #colors.append(viridis(i))
-        a = redblue((128+(subenr[i]*128)).astype(int)).tolist()
-        for j in range(len(a)):
-            a[j] = [1.,1.,1.,1.] if impv[j,i] > okpval else a[j]
-        data.append(a)
-    data = pd.DataFrame(data=data,columns=list(subenr.index),index= rand[sorting])
-    #data["clusters"]  = colors
-    
-    a = np.log2(1.01+merged[merged.columns[cols:annot]].iloc[rand].iloc[sorting].T)
-    if not precise:
-        for i in set(groups):
-            e = a[a.columns[subgroups[sorting]==i]].mean(1)
-            e = pd.DataFrame([e for i in range((subgroups[sorting]==i).sum())]).T
-            a[a.columns[subgroups[sorting]==i]] = e
-    
-    fig = sns.clustermap(a, vmin=vmin, vmax=vmax, figsize=size, z_score=0, colors_ratio=0.01, col_cluster=False,col_colors=data, xticklabels=False)
-    fig.ax_col_dendrogram.set_visible(False)
-    fig.fig.suptitle(title)
-    fig.savefig(folder + str(len(set(groups))) + '_clustermap_cobinding_enrichment_andrewplot.pdf')
-    plt.show()
