+++ setup.py
@@ -1,4 +1,10 @@
 from setuptools import setup
+import sys
+if sys.version_info.major < 3 or sys.version_info.minor < 2:
+    raise ValueError("genepy is only compatible with Python 3.3 and below")
+if sys.version_info.minor < 5:
+    import warnings
+    warnings.warn("genepy may not function properly on Python < 3.5")
 
 with open("README.md", 'r') as f:
     long_description = f.read()
@@ -11,8 +17,11 @@ setup(
     author='Jeremie Kalfon',
     author_email='jkobject@gmail.com',
     url="https://github.com/BroadInstitute/genepy",
-    packages=['cell_line_mapping-master/python'],  # same as name
+    packages=['genepy/cell_line_mapping-master/python',
+              'genepy/epigenetics/rose'],  # same as name
+    package_data={'genepy': ['data/*']},
     python_requires='>=3.5',
+    scripts=['genepy'],
     install_requires=[
         'rpy2-bioconductor-extensions',
         'gseapy',
@@ -44,6 +53,11 @@ setup(
         "typing",
         "venn",
         ],  # external packages as dependencies
+    classifiers=[
+        "Programming Language :: Python :: 3",
+        "Intended Audience :: Science/Research",
+        "Topic :: Scientific/Engineering :: Bio-Informatics",
+    ],
 )
 
 print("You might want to install Bowtie2, samtools, bwa and R to be able to use all functions of this package:\n\
